Restructured API files
