+++ backend/src/api2/api.py
@@ -20,9 +20,10 @@ from typing import (
 from sanic.log import logger
 
 import navi
-from base_types import InputId, OutputId
-from custom_types import NodeType, RunFn
-from node_check import (
+
+from .group import Group, GroupId, NestedGroup, NestedIdGroup
+from .input import BaseInput
+from .node_check import (
     NAME_CHECK_LEVEL,
     TYPE_CHECK_LEVEL,
     CheckFailedError,
@@ -30,10 +31,9 @@ from node_check import (
     check_naming_conventions,
     check_schema_types,
 )
-from nodes.base_input import BaseInput
-from nodes.base_output import BaseOutput
-from nodes.group import Group, GroupId, NestedGroup, NestedIdGroup
-from nodes.utils.exec_options import SettingsJson, get_execution_options
+from .output import BaseOutput
+from .settings import SettingsJson, get_execution_options
+from .types import InputId, NodeType, OutputId, RunFn
 
 KB = 1024**1
 MB = 1024**2
