--- backend/src/base_types.py
@@ -1,5 +1,0 @@
-from typing import NewType
-
-NodeId = NewType("NodeId", str)
-InputId = NewType("InputId", int)
-OutputId = NewType("OutputId", int)
