+++ backend/src/nodes/properties/inputs/__init__.py
@@ -1,5 +1,3 @@
-from nodes.base_input import *
-
 from .file_inputs import *
 from .generic_inputs import *
 from .image_dropdown_inputs import *
