+++ backend/src/nodes/properties/inputs/file_inputs.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import os
 from typing import Literal, Union
 
-from nodes.base_input import BaseInput
+from api2 import BaseInput
 
 # pylint: disable=relative-beyond-top-level
 from ...impl.image_formats import get_available_image_formats
