+++ backend/src/nodes/properties/inputs/ncnn_inputs.py
@@ -1,4 +1,4 @@
-from nodes.base_input import BaseInput
+from api2 import BaseInput
 
 from ...impl.ncnn.model import NcnnModelWrapper
