+++ backend/src/nodes/properties/inputs/onnx_inputs.py
@@ -1,5 +1,5 @@
 import navi
-from nodes.base_input import BaseInput
+from api2 import BaseInput
 
 from ...impl.onnx.model import OnnxModel, OnnxModels, OnnxRemBg, is_rembg_model
 from .generic_inputs import DropDownInput
