+++ backend/src/nodes/properties/outputs/__init__.py
@@ -1,5 +1,3 @@
-from nodes.base_output import *
-
 from .file_outputs import *
 from .generic_outputs import *
 from .numpy_outputs import *
