+++ backend/src/nodes/properties/outputs/generic_outputs.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import Union
 
 import navi
-from nodes.base_output import BaseOutput
+from api2 import BaseOutput
 
 from ...impl.color.color import Color
 from ...utils.format import format_color_with_channels
