+++ backend/src/nodes/properties/outputs/onnx_outputs.py
@@ -1,5 +1,5 @@
 import navi
-from nodes.base_output import BaseOutput
+from api2 import BaseOutput
 
 from ...impl.onnx.model import OnnxModel
