+++ backend/src/nodes/properties/outputs/pytorch_outputs.py
@@ -1,7 +1,7 @@
 from typing import List
 
 import navi
-from nodes.base_output import BaseOutput, OutputKind
+from api2 import BaseOutput, OutputKind
 
 from ...impl.pytorch.types import PyTorchModel
 from ...utils.format import format_channel_numbers
