+++ backend/src/packages/chaiNNer_onnx/__init__.py
@@ -1,6 +1,6 @@
 from sanic.log import logger
 
-from api import KB, MB, Dependency, add_package
+from api2 import KB, MB, Dependency, add_package
 from gpu import nvidia_is_available
 from system import is_arm_mac
