+++ backend/src/packages/chaiNNer_standard/image/batch_processing/merge_spritesheet.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import numpy as np
 
-from api import Collector, IteratorInputInfo
+from api2import Collector, IteratorInputInfo
 from nodes.properties.inputs import ImageInput, NumberInput
 from nodes.properties.outputs import ImageOutput
